requirements.txt
streamlit~=1.39.0
plotly~=5.24.1
pandas~=2.2.3
pyowm~=3.3.0
huggingface-hub~=0.26.0
python-dotenv~=1.0.1